sonar-fix

+++ webapp/src/main/java/com/getjavajob/training/web1803/webapp/servlets/MessageServlet.java
@@ -39,25 +39,25 @@ public class MessageServlet extends HttpServlet {
                 groupId = assignId;
                 location = "group.jsp";
                 break;
+            default:
+                location = "";
+                break;
         }
-        switch (action) {
-            case "new":
-                String text = request.getParameter("inputNewMessage");
-                InputStream photo = null;
-                String photoFileName = null;
-                Part filePart = request.getPart("uploadImage");
-                if (filePart != null) {
-                    photo = filePart.getInputStream();
-                    photoFileName = filePart.getSubmittedFileName();
-                }
-                HttpSession session = request.getSession();
-                int currentId = (Integer) session.getAttribute("id");
-                service.create(groupId, accountId, type, photo, photoFileName, text, new SimpleDateFormat("yyyy-MM-dd").format(new Date()), currentId);
-                break;
-            case "remove":
-                int messageId = Integer.valueOf(request.getParameter("messageId"));
-                service.remove(messageId);
-                break;
+        if (action.equals("new")) {
+            String text = request.getParameter("inputNewMessage");
+            InputStream photo = null;
+            String photoFileName = null;
+            Part filePart = request.getPart("uploadImage");
+            if (filePart != null) {
+                photo = filePart.getInputStream();
+                photoFileName = filePart.getSubmittedFileName();
+            }
+            HttpSession session = request.getSession();
+            int currentId = (Integer) session.getAttribute("id");
+            service.create(groupId, accountId, type, photo, photoFileName, text, new SimpleDateFormat("yyyy-MM-dd").format(new Date()), currentId);
+        } else if (action.equals("remove")) {
+            int messageId = Integer.valueOf(request.getParameter("messageId"));
+            service.remove(messageId);
         }
         response.sendRedirect(location + "?id=" + assignId);
     }
